perf(frontend): make the first press of Start paint immediately
Picking a school and pressing "i'm ready to start" paused before the
setup card appeared, but only ever on the first press. Everything the
second press reused, the first press had to fetch, with nothing on
screen while it did.

/onboarding is force-dynamic, and Next prefetches a dynamic route only
as far as its nearest loading boundary. There was no loading.tsx, so
there was no boundary: measured on a production build, the prefetch the
hero's Link fires returned 186 bytes and named zero chunks to preload.
The click then paid for a cold RSC round trip plus 185 kB raw / ~57 kB
gzipped of JavaScript the browser had never seen, about 40 kB gzipped of
which was the Firebase Auth SDK doing nothing until "Text me a code".
The page also awaited getSession() and getUserByAuthUid() above its own
Suspense boundary, so nothing streamed until both had answered.

Three changes:

  loading.tsx            gives the prefetcher a boundary to stop at. The
                         same prefetch now returns 10.3 kB of real card
                         markup and preloads five chunks, so the route's
                         JS lands while the student is still choosing a
                         school.
  page.tsx               synchronous again; an async child does the two
                         reads below the boundary, so the card flushes
                         first and the wizard arrives after.
  firebaseClient.ts      firebase/app and firebase/auth move to a dynamic
                         import, warmed from an effect once the wizard
                         has mounted rather than parsed before it can
                         hydrate.

Shell, Progress and the page frame move to components/onboarding/shell.tsx
so loading.tsx can render the exact same card on the server. That file
must stay free of hooks: a "use client" on it drags the whole wizard
bundle into the loading boundary and silently undoes this.

  prefetch payload      186 B, 0 chunks  ->  10.3 kB, 5 chunks
  new JS on the hop     185 kB / 57 gz   ->  69 kB / 21 gz
  route JS              56.5 kB          ->  16.6 kB
  first load JS         167 kB           ->  132 kB

Rationale and the rules this leaves behind: docs/design/16.
